refactor layout structure
