+++ src/App.tsx
@@ -1,28 +1,21 @@
-import React from 'react';
-import './App.css';
-import './components/section/sectionMain/style.css'
-import './components/section/sectionAbout/style.css'
-import './components/section/sectionSlider/style.css'
-import './components/section/sectionHelp/style.css'
-import './components/section/sectionDonation/style.css'
-import './components/header/style.css'
-import './components/footer/style.css'
-import { SectionMain } from './components/section/sectionMain/SectionMain';
-import { SectionAbout } from './components/section/sectionAbout/SectionAbout';
-import { SectionSlider } from './components/section/sectionSlider/SectionSlider';
-import { SectionHelp } from './components/section/sectionHelp/SectionHelp';
-import { SectionDonation } from './components/section/sectionDonation/SectionDonation';
-import { Footer } from './components/footer/Footer';
+import React from "react";
+import "./App.css";
+import "./pages/mainPage/sectionMain/style.css";
+import "./pages/mainPage/sectionAbout/style.css";
+import "./pages/mainPage/sectionSlider/style.css";
+import "./pages/mainPage/sectionHelp/style.css";
+import "./pages/mainPage/sectionDonation/style.css";
+import "./components/portal/header.css";
+import "./components/portal/footer.css";
+import { Portal } from "./components/portal/Portal";
+import { MainPage } from "./pages/mainPage/MainPage";
 
 function App() {
   return (
     <div className="App">
-      <SectionMain />
-      <SectionAbout />
-      <SectionSlider />
-      <SectionHelp />
-      <SectionDonation />
-      <Footer />
+      <Portal>
+        <MainPage />
+      </Portal>
     </div>
   );
 }
