--- src/components/header/Header.tsx
@@ -1,16 +1,0 @@
-export const Header = function () {
-    return (
-        <header>
-        <div className="logo">
-          <div className="logo-title">Cozy House</div>
-          <div className="logo-subtitle">Shelter for pets in Boston</div>
-        </div>
-        <nav>
-          <div className="active">About the shelter</div>
-          <div>Our pets</div>
-          <div>Help the shelter</div>
-          <div>Contacts</div>
-        </nav>
-      </header>
-    )
-}
