+++ src/pages/mainPage/sectionAbout/SectionAbout.tsx
