+++ src/pages/mainPage/sectionAbout/style.css
