+++ src/pages/mainPage/sectionDonation/style.css
