+++ src/pages/mainPage/sectionHelp/SectionHelp.tsx
