+++ src/pages/mainPage/sectionMain/style.css
