+++ src/pages/mainPage/sectionSlider/SectionSlider.tsx
