+++ src/pages/mainPage/sectionSlider/style.css
