+++ tsconfig.json
@@ -22,5 +22,5 @@
   },
   "include": [
     "src"
-  ]
+, "src/pages/mainPage/section"  ]
 }
